Changes Symphony production server name.
The new Symphony server will be renamed to bodoni when it is put into
production. In order for the pull* scripts to work, they should be
getting data from bodoni and not bodoni-local. Using bodoni-local was
put in place during the last Symphony server migration and is no longer
needed.